# src/Harri.SchoolDemoAPI/wwwroot/openapi-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "description" : "Demo API about students, schools, and students applications to schools",
    "title" : "Swagger School ADMIN - OpenAPI 3.0",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
...
  "paths" : {
    "/student/{sId}" : {
      "delete" : {
        "description" : "Delete a student by sId",
        "operationId" : "deleteStudent",
        "parameters" : [ {
          "description" : "ID of student to delete",
          "explode" : false,
          "in" : "path",
          "name" : "sId",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/StudentId"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "409" : {
            "description" : "Conflict. Student has applications which must be deleted first"
          },
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "Student not found"
          }
        },
        "summary" : "Delete a student",
        "tags" : [ "student" ]
      },
      "get" : {
        "description" : "Get a student by sId",
        "operationId" : "getStudent",
        "parameters" : [ {
          "description" : "ID of student to return",
          "explode" : false,
          "in" : "path",
          "name" : "sId",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/StudentId"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Student"
                }
              }
            },
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "Student not found"
          }
        },
        "summary" : "Get a student",
        "tags" : [ "student" ]
      }
    },
    "/student" : {
      "post" : {
        "description" : "Add a new student",
        "operationId" : "addStudent",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewStudent"
              }
            }
          },
          "description" : "Create a new student",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StudentId"
                }
              }
            },
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Invalid input"
          }
        },
        "summary" : "Add a new student",
        "tags" : [ "student" ]
      },
      "put" : {
        "description" : "Update an existing student by Id",
        "operationId" : "updateStudent",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Student"
              }
            }
          },
          "description" : "Update an existent student",
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "Student not found"
          }
        },
        "summary" : "Update an existing student",
        "tags" : [ "student" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Student" : {
        "example" : {
          "name" : "Garry Peterson",
          "GPA" : 3.9,
          "sId" : 1234
        },
        "properties" : {
          "sId" : {
            "example" : 1234,
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "example" : "Garry Peterson",
            "type" : "string"
          },
          "GPA" : {
            "example" : 3.9,
            "type" : "number"
          }
        },
        "required" : [ "sId" ],
        "type" : "object"
      },
      "NewStudent" : {
        "example" : {
          "name" : "Garry Peterson",
          "GPA" : 3.9
        },
        "properties" : {
          "name" : {
            "example" : "Garry Peterson",
            "type" : "string"
          },
          "GPA" : {
            "example" : 3.9,
            "type" : "number"
          }
        },
        "type" : "object"
      },
...
      "StudentId" : {
        "example" : 1234,
        "format" : "int64",
        "type" : "integer"
      },
...
    }
  }
}